refactor lite and fuse awq and smooth_quant for vision model
