+++ lmdeploy/lite/calibration/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .awq_calibration import AWQCalibrationContext
+from .calibration import CalibrationContext
+from .vision_calibration import VisionCalibrationContext
+
+__all__ = [
+    'CalibrationContext', 'AWQCalibrationContext', 'VisionCalibrationContext'
+]
