+++ lmdeploy/lite/quantization/__init__.py
@@ -1,9 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .activation import ActivationObserver, KVCacheObserver
-from .calibration import CalibrationContext, CalibrationContextV2
+from .activation import ActivationObserver
 from .weight import WeightQuantizer
 
-__all__ = [
-    'WeightQuantizer', 'ActivationObserver', 'KVCacheObserver',
-    'CalibrationContext', 'CalibrationContextV2'
-]
+__all__ = ['WeightQuantizer', 'ActivationObserver']
