+++ lmdeploy/lite/quantization/activation/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .observer import ActivationObserver, KVCacheObserver
+from .observer import ActivationObserver
 
-__all__ = ['ActivationObserver', 'KVCacheObserver']
+__all__ = ['ActivationObserver']
