+++ lmdeploy/vl/model/internvl.py
@@ -93,9 +93,11 @@ class InternVLVisionModel(VisonModel):
             else:
                 self.vl_model = model
             model.half()
-        from lmdeploy.pytorch.modeling.convert_to_qmodules import \
-            convert_to_qmodules
-        convert_to_qmodules(model)
+        if getattr(config.vision_config, 'quantization_config',
+                   None) is not None:
+            from lmdeploy.pytorch.modeling.convert_to_qmodules import \
+                convert_to_qmodules
+            convert_to_qmodules(model)
 
         from accelerate import load_checkpoint_and_dispatch
         with disable_logging():
